Multiround aggregator added and mean refactored
